+++ client/src/pages/Profile/Profile.jsx
@@ -6,7 +6,7 @@ import Account from './Account/Account'
 import Orders from './Orders/Orders'
 
 import AdminDashboard from './AdminDashboard/AdminDashboard'
-import styles from './Profile.module.css'
+
 
 import logo from '../../assets/icons/air_land-black.svg'
 import styles from './Profile.module.css'
